Refactor chat interface: remove old frontend files and implement new ChatModal component with enhanced styling and functionality
