+++ Backend/package.json
@@ -14,6 +14,7 @@
     "dependencies": {
         "@clerk/clerk-sdk-node": "^4.13.23",
         "@clerk/express": "^1.5.0",
+        "@google/generative-ai": "^0.24.1",
         "@libsql/client": "^0.15.7",
         "cors": "^2.8.5",
         "dotenv": "^16.5.0",
