+++ Backend/src/routes/empresas.js
@@ -15,7 +15,8 @@ router.post("/", async (req, res) => {
         latitud,
         longitud,
         tipo,
-    } = req.body;    try {
+    } = req.body;
+    try {
         // Primero, intentamos crear una tabla de seguimiento de IDs si no existe
         try {
             await db.execute({
@@ -23,45 +24,45 @@ router.post("/", async (req, res) => {
                     tipo TEXT NOT NULL,
                     ultimo_numero INTEGER NOT NULL,
                     PRIMARY KEY (tipo)
-                )`
+                )`,
             });
         } catch (createErr) {
             console.error("Error al crear tabla de seguimiento:", createErr);
             // Continuamos con la ejecución aunque haya fallado la creación
         }
-        
+
         // Obtener el número más alto utilizado para empresas
         let nextNum = 1;
-        
+
         // Primero verificamos en la tabla de seguimiento
         const trackingResult = await db.execute({
-            sql: "SELECT ultimo_numero FROM id_tracking WHERE tipo = 'empresa'"
+            sql: "SELECT ultimo_numero FROM id_tracking WHERE tipo = 'empresa'",
         });
-        
+
         if (trackingResult.rows.length > 0) {
             // Ya hay un registro de seguimiento, usamos ese número + 1
             nextNum = trackingResult.rows[0].ultimo_numero + 1;
         } else {
             // No hay registro en la tabla de seguimiento, buscamos en la tabla empresas
             const lastIdResult = await db.execute({
-                sql: "SELECT id FROM empresas WHERE id LIKE 'emp%' ORDER BY CAST(SUBSTR(id, 4) AS INTEGER) DESC LIMIT 1"
+                sql: "SELECT id FROM empresas WHERE id LIKE 'emp%' ORDER BY CAST(SUBSTR(id, 4) AS INTEGER) DESC LIMIT 1",
             });
-            
+
             if (lastIdResult.rows.length > 0) {
                 const lastId = lastIdResult.rows[0].id;
-                const lastNum = parseInt(lastId.replace('emp', ''));
+                const lastNum = parseInt(lastId.replace("emp", ""));
                 nextNum = lastNum + 1;
             }
-            
+
             // Insertamos un registro inicial en la tabla de seguimiento
             await db.execute({
                 sql: "INSERT INTO id_tracking (tipo, ultimo_numero) VALUES (?, ?)",
-                args: ['empresa', nextNum]
+                args: ["empresa", nextNum],
             });
         }
-        
+
         const id = `emp${nextNum}`;
-          await db.execute({
+        await db.execute({
             sql: `INSERT INTO empresas 
         (id, nombre, descripcion, email, telefono, sitio_web, direccion, latitud, longitud, tipo)
         VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
@@ -78,11 +79,11 @@ router.post("/", async (req, res) => {
                 tipo,
             ],
         });
-        
+
         // Actualizamos el contador en la tabla de seguimiento
         await db.execute({
             sql: "UPDATE id_tracking SET ultimo_numero = ? WHERE tipo = ?",
-            args: [nextNum, 'empresa']
+            args: [nextNum, "empresa"],
         });
 
         res.status(201).json({ id, nombre });
@@ -180,48 +181,48 @@ router.delete("/:id", async (req, res) => {
             sql: "SELECT id FROM atractivos WHERE empresa_id = ?",
             args: [id],
         });
-        
+
         // Para cada atractivo, eliminar registros relacionados
         for (const atractivo of atractivos.rows) {
             const atractivoId = atractivo.id;
-            
+
             // Eliminar referencias en rutas_atractivos
             await db.execute({
                 sql: "DELETE FROM rutas_atractivos WHERE atractivo_id = ?",
                 args: [atractivoId],
             });
-            
+
             // Eliminar reseñas relacionadas
             await db.execute({
                 sql: "DELETE FROM reseñas WHERE atractivo_id = ?",
                 args: [atractivoId],
             });
-            
+
             // Eliminar favoritos relacionados
             await db.execute({
                 sql: "DELETE FROM favoritos WHERE atractivo_id = ?",
                 args: [atractivoId],
             });
-            
+
             // Eliminar imágenes relacionadas con el atractivo
             await db.execute({
                 sql: "DELETE FROM imagenes WHERE entidad_tipo = 'atractivo' AND entidad_id = ?",
                 args: [atractivoId],
             });
-            
+
             // Finalmente eliminar el atractivo
             await db.execute({
                 sql: "DELETE FROM atractivos WHERE id = ?",
                 args: [atractivoId],
             });
         }
-        
+
         // Eliminar imágenes relacionadas con la empresa
         await db.execute({
             sql: "DELETE FROM imagenes WHERE entidad_tipo = 'empresa' AND entidad_id = ?",
             args: [id],
         });
-        
+
         // Finalmente, eliminar la empresa
         const result = await db.execute({
             sql: "DELETE FROM empresas WHERE id = ?",
@@ -232,7 +233,9 @@ router.delete("/:id", async (req, res) => {
             return res.status(404).json({ error: "Empresa no encontrada" });
         }
 
-        res.json({ mensaje: "Empresa y sus atractivos eliminados exitosamente" });
+        res.json({
+            mensaje: "Empresa y sus atractivos eliminados exitosamente",
+        });
     } catch (err) {
         console.error("Error al eliminar empresa:", err);
         res.status(500).json({ error: "Error al eliminar empresa" });
@@ -240,7 +243,7 @@ router.delete("/:id", async (req, res) => {
 });
 /**
  * @swagger
- * /empresas:
+ * /api/empresas:
  *   post:
  *     summary: Crea una nueva empresa
  *     tags: [Empresas]
@@ -297,7 +300,7 @@ router.delete("/:id", async (req, res) => {
 
 /**
  * @swagger
- * /empresas:
+ * /api/empresas:
  *   get:
  *     summary: Obtiene la lista de todas las empresas
  *     tags: [Empresas]
@@ -314,7 +317,7 @@ router.delete("/:id", async (req, res) => {
 
 /**
  * @swagger
- * /empresas/{id}:
+ * /api/empresas/{id}:
  *   get:
  *     summary: Obtiene una empresa por su ID
  *     tags: [Empresas]
@@ -338,7 +341,7 @@ router.delete("/:id", async (req, res) => {
 
 /**
  * @swagger
- * /empresas/{id}:
+ * /api/empresas/{id}:
  *   put:
  *     summary: Actualiza una empresa por su ID
  *     tags: [Empresas]
@@ -385,7 +388,7 @@ router.delete("/:id", async (req, res) => {
 
 /**
  * @swagger
- * /empresas/{id}:
+ * /api/empresas/{id}:
  *   delete:
  *     summary: Elimina una empresa por su ID
  *     tags: [Empresas]
