--- Chat/frontend/astro.config.mjs
@@ -1,5 +1,0 @@
-// @ts-check
-import { defineConfig } from 'astro/config';
-
-// https://astro.build/config
-export default defineConfig({});
