--- Chat/frontend/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "frontend",
-  "type": "module",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "astro dev",
-    "build": "astro build",
-    "preview": "astro preview",
-    "astro": "astro"
-  },
-  "dependencies": {
-    "astro": "^5.8.0",
-    "axios": "^1.9.0"
-  }
-}
