--- Chat/frontend/src/components/ChatStyles.astro
@@ -1,128 +1,0 @@
----
-
----
-
-<style is:global>
-    body {
-        margin: 0;
-        padding: 0;
-        font-family: Arial, sans-serif;
-        background: #121212;
-        color: #fff;
-    }
-
-    .chat-container {
-        max-width: 500px;
-        width: 500px;
-        margin: 20px auto;
-        height: 50vh;
-        display: flex;
-        flex-direction: column;
-        background: #1e1e1e;
-        border-radius: 10px;
-        box-shadow: 0 2px 10px rgba(0, 0, 0, 0.5);
-    }
-
-    .messages {
-        flex-grow: 1;
-        overflow-y: auto;
-        padding: 20px;
-        display: flex;
-        flex-direction: column;
-    }
-
-    .message {
-        max-width: 65%;
-        margin: 8px;
-        padding: 12px 18px;
-        border-radius: 18px;
-        line-height: 1.4;
-        position: relative;
-        word-wrap: break-word;
-        animation: fadeIn 0.3s ease-in-out;
-    }
-
-    @keyframes fadeIn {
-        from {
-            opacity: 0;
-            transform: translateY(10px);
-        }
-        to {
-            opacity: 1;
-            transform: translateY(0);
-        }
-    }
-
-    .message.user {
-        background: linear-gradient(135deg, #00b2ff, #0084ff);
-        color: white;
-        align-self: flex-end;
-        border-bottom-right-radius: 4px;
-        box-shadow: 0 1px 2px rgba(0, 0, 0, 0.2);
-        padding: 1px 10px;
-    }
-
-    .message.assistant {
-        background: #333;
-        color: #eee;
-        align-self: flex-start;
-        border-bottom-left-radius: 4px;
-        box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
-        text-align: left;
-    }
-
-    .error {
-        background: linear-gradient(135deg, #ff4444, #cc0000);
-        color: white;
-        align-self: flex-start;
-        border-bottom-left-radius: 4px;
-    }
-
-    .input-form {
-        display: flex;
-        padding: 20px;
-        gap: 10px;
-        border-top: 1px solid #333;
-        background: #222;
-    }
-
-    input {
-        flex-grow: 1;
-        padding: 12px 20px;
-        border: 2px solid #555;
-        border-radius: 25px;
-        outline: none;
-        font-size: 16px;
-        transition: border-color 0.3s ease;
-        background: #333;
-        color: #fff;
-    }
-
-    input:focus {
-        border-color: #0084ff;
-    }
-
-    button {
-        padding: 12px 28px;
-        background: linear-gradient(135deg, #00b2ff, #0084ff);
-        color: white;
-        border: none;
-        border-radius: 25px;
-        cursor: pointer;
-        transition: all 0.3s ease;
-        font-weight: bold;
-        font-size: 16px;
-        box-shadow: 0 2px 4px rgba(0, 130, 255, 0.2);
-    }
-
-    button:hover {
-        background: linear-gradient(135deg, #0084ff, #0066cc);
-        transform: translateY(-1px);
-        box-shadow: 0 4px 8px rgba(0, 130, 255, 0.3);
-    }
-
-    button:active {
-        transform: translateY(0);
-        box-shadow: 0 2px 4px rgba(0, 130, 255, 0.2);
-    }
-</style>
