--- Chat/frontend/src/pages/index.astro
@@ -1,91 +1,0 @@
----
-import ChatScript from "../components/ChatScript.astro";
-import ChatStyles from "../components/ChatStyles.astro";
----
-
-<html lang="es">
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width" />
-        <title>Chatturify</title>
-        <script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"
-        ></script>
-        <script
-            src="https://cdn.jsdelivr.net/npm/dompurify@3.0.6/dist/purify.min.js"
-        ></script>
-    </head>
-    <body>
-        <main>
-            <section class="hero">
-                <h1>Chatturify</h1>
-                <p>
-                    ¡Bienvenido a Chatturify! Un chat para hablar con un
-                    asistente.
-                </p>
-            </section>
-
-            <div class="chat-container" id="chat-container">
-                <div class="messages" id="messages">
-                    <div class="message assistant">
-                        Hola! ¿En qué puedo ayudarte?
-                    </div>
-                </div>
-
-                <form id="chat-form" class="input-form">
-                    <input
-                        type="text"
-                        id="user-input"
-                        placeholder="Escribe tu mensaje..."
-                        required
-                    />
-                    <button type="submit">Enviar</button>
-                </form>
-            </div>
-        </main>
-
-        <ChatScript />
-        <ChatStyles />
-    </body>
-    <style>
-        main {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            height: 100vh;
-            text-align: center;
-            padding: 20px;
-        }
-
-        .hero {
-            max-width: 600px;
-        }
-
-        h1 {
-            font-size: 3em;
-            margin-bottom: 20px;
-        }
-
-        p {
-            font-size: 1.2em;
-            color: #ccc;
-        }
-
-        .chat-bubble {
-            position: fixed;
-            bottom: 20px;
-            right: 20px;
-            background: linear-gradient(135deg, #00b2ff, #0084ff);
-            color: white;
-            width: 60px;
-            height: 60px;
-            border-radius: 50%;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            text-decoration: none;
-            font-size: 1.2em;
-            box-shadow: 0 2px 5px rgba(0, 0, 0, 0.3);
-        }
-    </style>
-</html>
