+++ Frontend/src/layouts/MainLayout.astro
@@ -1,6 +1,7 @@
 ---
 import Navbar from "../components/Navbar.astro";
 import Footer from "../components/Footer.astro";
+import ChatModal from "../components/ChatModal.astro";
 ---
 
 <html lang="es">
@@ -14,5 +15,6 @@ import Footer from "../components/Footer.astro";
             <slot />
         </main>
         <Footer />
+        <ChatModal />
     </body>
 </html>
